refactor(web): remove trailing space in "ClipTabs .tsx"
