+++ service/vspo-schedule/web/src/components/Templates/ClipList.tsx
