+++ service/vspo-schedule/web/src/components/Templates/index.ts
@@ -1,4 +1,4 @@
-export * from "./ClipList ";
+export * from "./ClipList";
 export * from "./ClipTabs";
 export * from "./Livestreams";
 export * from "./RelatedVideos";
